+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: b79ed685f78b8d5356795eb2d1112f830abaf58e
+// Checksum: edc35ebb9a15ca000a4026b86603f92d60f6bf7a
 //
 // ignore_for_file: unused_import
 
@@ -4235,13 +4235,13 @@ final class _EvaluateVisitor
     required bool sourceMap,
     bool warnForColor = false,
   }) {
-    var targetLocations = sourceMap ? <SourceLocation>[] : null;
+    var targetOffsets = sourceMap ? <int>[] : null;
     var oldInSupportsDeclaration = _inSupportsDeclaration;
     _inSupportsDeclaration = false;
     var buffer = StringBuffer();
     var first = true;
     for (var value in interpolation.contents) {
-      if (!first) targetLocations?.add(SourceLocation(buffer.length));
+      if (!first) targetOffsets?.add(buffer.length);
       first = false;
 
       if (value is String) {
@@ -4279,8 +4279,8 @@ final class _EvaluateVisitor
 
     return (
       buffer.toString(),
-      targetLocations.andThen(
-        (targetLocations) => InterpolationMap(interpolation, targetLocations),
+      targetOffsets.andThen(
+        (targetOffsets) => InterpolationMap(interpolation, targetOffsets),
       ),
     );
   }