+++ pkg/sass-parser/test/utils.ts
@@ -4,12 +4,17 @@
 
 import {
   AnyExpression,
+  AnySimpleSelector,
   ChildNode,
   ChildProps,
+  CompoundSelector,
   ExpressionProps,
   GenericAtRule,
   Interpolation,
   Root,
+  Rule,
+  SelectorList,
+  SimpleSelectorProps,
   scss,
 } from '../lib';
 
@@ -22,6 +27,28 @@ export function parseExpression<T extends AnyExpression>(text: string): T {
   return expression;
 }
 
+/** Parses selector list from {@link text}. */
+export function parseSelector(text: string): SelectorList {
+  const rule = scss.parse(`${text} {}`).nodes[0] as Rule;
+  expect(rule.type).toEqual('rule');
+  return rule.parsedSelector;
+}
+
+/** Parses simple selector from {@link text}. */
+export function parseSimpleSelector<T extends AnySimpleSelector>(
+  text: string,
+): T {
+  const rule = scss.parse(`${text} {}`).nodes[0] as Rule;
+  expect(rule.type).toEqual('rule');
+  expect(rule.parsedSelector.nodes).toHaveLength(1);
+  const complex = rule.parsedSelector.nodes[0];
+  expect(complex.nodes).toHaveLength(1);
+  const component = complex.nodes[0];
+  expect(component.combinator).toBeUndefined();
+  expect(component.compound.nodes).toHaveLength(1);
+  return component.compound.nodes[0] as T;
+}
+
 /** Constructs a new node from {@link props} as in child node injection. */
 export function fromChildProps<T extends ChildNode>(props: ChildProps): T {
   return new Root({nodes: [props]}).nodes[0] as T;
@@ -32,4 +59,11 @@ export function fromExpressionProps<T extends AnyExpression>(
   props: ExpressionProps,
 ): T {
   return new Interpolation({nodes: [props]}).nodes[0] as T;
+}
+
+/** Constructs a new simple selector from {@link props}. */
+export function fromSimpleSelectorProps<T extends AnySimpleSelector>(
+  props: SimpleSelectorProps,
+): T {
+  return new CompoundSelector({nodes: [props]}).nodes[0] as T;
 }